Improve management of Supervisor from playbook files
Use the Ansible supervisorctl module, instead of directly executing
"supervisorctl" via the command or shell modules (in places where it
was still done). This replaces a task that caused a syntax warning:

     [WARNING]: when statements should not include jinja2 templating
    delimiters such as {{ }} or {% %}. Found:
    supervisor_applications.stdout.find('{{ celery_application_name
    }}') != -1

Supervisor is installed in the base role, so it does not need to be
installed in other roles that use it. However, add a task in these
roles to ensure the Supervisor service is running.

Since handlers already exist in the playbook to restart applications
controlled by Supervisor after a configuration change, this does not
need to be done separately from setup_supervisor.yml playbook files.