Move SSL certificate/key into variables file for each environment
The domain name of the nginx server is likely to be different when
running vagrant.yml versus the other playbooks. Move the "ssl_crt"
and "ssl_key" settings to the variables file for each environment.

For development, change the provided SSL certificate's common name
to "dev.example.com" (the machine hostname in the inventory file).
While this hostname and SSL certificate/key need to be customized
to use the development playbook anyway, this is more illustrative.

Remove an unneeded override for the "nginx_server_name" variable
which does not change it from its default value.